Style playground a bit
