+++ packages/docs/package.json
@@ -21,6 +21,7 @@
     "fumadocs-twoslash": "^1.1.3",
     "fumadocs-ui": "13.4.10",
     "next": "^14.2.12",
+    "next-themes": "^0.3.0",
     "next-yak": "workspace:^",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
